requirements.txt
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
hachoir==3.3.0
idna==3.10
mutagen==1.47.0
peewee==3.17.6
psutil==6.0.0
pycryptodomex==3.20.0
pywin32==306
PyYAML==6.0.2
requests==2.32.3
urllib3==2.2.3
websockets==13.1
youtube-dl==2021.12.17
yt-dlp==2024.9.27
json5==0.10.0
Flask==3.1.0
pillow<11.0
moviepy==2.1.2
selenium==4.28.1
beautifulsoup4==4.12.3 
wget==3.2
asyncio==3.4.3
